Frontend/requirements.txt
customtkinter
opencv-python
Pillow
picamera2
pyserial